XD-2195 Make XD cache break when version bump up
The purpose of this change is to solve the problem that browser
cache XD's static asssets so new file change didn't get applied.

The basic idea is to append a version based suffix to all css/js path
in html files to break the cache.

This commit includes the following changes to make it work:
1. introduce grunt-cache-breaker to append version suffix to css/js path
2. Add Version.js and make it the single source of trunth of a version number
3. add random number to icomoon related url in icon.less to break the cache

Change-Id: I5540161776fd47cda6a501092def83beb63d4954