ENG-4808 - MFS support for Xcalar SaaS
This modifies the passport-cognito strategy to accept two types of
inputs:
- username and password
- idToken, accessToken, refreshToken

The username and password deals with a non-MFA case where a username
and password is sufficient for authentication.

The three tokens are used for a MFA case.  There, the User Pool
authentication demands that the MFA auth information is captured
and sent mid-authentication flow.   This should not happen in the
middle of a Passport strategy.  Instead, the User Pool auth is
presumed to occur before the strategy is invoked.  The tokens
presented by the User Pool are verified and augmented with
Identity Pool data.

Change-Id: I5b7c6913bbebe688cfe7e65f7f2f2f6ce20dd35c