ENG-2841, ENG-2936 Use url to pass in sessionId from cloudLogin to XD
Because cloudLogin and XD are in different domain, this commit
use the url way to pass the necessary info from cloudLogin
to XD in order to make the whole login flow work.

This commit also includes the integration code on XD's login.js
to call login api using the sessionId

This commit also add some alert/error message for cloudLogin page.

There are some hack codes check in with the commit, which will be
cleaned up after the full integration works

Change-Id: I5290f6ec32c3d1c7c2d5d3008f06ef1eee2accdf