ENG-4752 Confluent Kafka sink connector
"assets/misc/cfconnector" is the code for Confluent Kafka Connector.

It is not a part of either XD nor XCE, so it's excluded from build
process. To build it, we need to manually run "mvn clean package".

We keep it in XD repo for now because the XCE test runs too long.

Change-Id: I5c5a816babdbbf78c89d49ff23426c1f5c2cb29a