ENGR-5040 - Add package-lock.json to other expServer and tests that require node modules.
The ostensible purpose of this change was to add package-lock.json files
to two directories (service/expServer and assets/test/integrationTest/xcrpc) and
then adjust the build to be more compatible the change.

This was addressed, but a number of other problems were discovered that were also
fixed.  The Makefile needed to be adjusted to deal with three issues:

1) The package-lock.json [$(PACKAGE_LOCK)] make target was creating the file, populating
node_modules, and then creating a tar file that could be re-used later.  This was broken
and producing a tar file named node_modules-.tar that could not be reused later.  As far
as I can tell, this is because the package_tar make variable was being treated as if it
was a procedure call (re-runs every time) but make does not treat it that way.  Make
treats it like a variable -- it gets evaluated once per make target and does not get
re-evaluated within the same target.  So, if package-lock.json does not exist, package_tar
becomes node_modules-.tar for that make target and stays that way even if the
package-lock.json is created during the target.

SOLUTION: the package-lock.json file just creates package-lock.json and node_modules
without archiving them.

2) Npm likes to make somewhat benign changes to the package-lock.json file that do not
affect the package put in node_modules, but does affect the checksum of the file.  So,
if the node_modules/.tstamp target ended up running npm install, a package-lock.json
file with a different checksum would exist that does not match the tarfile produced
by the target.  That change would not be detected after npm install is run because
of the problem in 1).  Subsequent runs of the Makefile would not find that cached tarfile.

SOLUTION: the node_modules/.tstamp runs `npm install --package-lock-only --save-dev`
to update the package-lock.json file and its checksum before checking about tarfiles.

3) Trivially, the package-lock.json [$(PACKAGE_LOCK)] make target was using $(@D) to
create an empty directory.

SOLUTION: $(@D) is removed.

The Gruntfile needed to be modified in two places to adjust to using package-lock.json
This is because the expServer and some tests were using the following npm commands:

1. npm install --no-save
2. npm uninstall --no-save xcalar xcalarsdk
3. npm upgrade --no-save

Command #1 ensures that the installation is clean and whole.  Command #2 removes
two Xcalar-generated packages for replacement by newly-generated code.  Command #3
causes the newly-generated code to be installed.  UNFORTUNATELY, it has the side
effect of upgrading every package in node_modules, which is potentially very bad and
causes the sorts of problems that package-lock.json is supposed to fix.

SOLUTION: Replace command #3 with this:

3. npm install xcalar-0.0.1.tgz xcalarsdk-0.0.1.tgz

This forces the newly generated code to be installed, and all other packages
remain more or less as they are.

Finally, the Gruntfile patches a number of Grunt plugins in order to get the build
properties we want.  One of the patches alters the grunt-assets-inline plugin to
use a dependency called uglify-es instead of one called uglify-js; uglify-js does
not uglify the code standard Xcalar uses and uglify-es does with a compatible
interface.  This patch WAS modifying package.json to reflect the dependency
change.  That, again, resulted in a package-lock.json file that would not match
any existing cached tar file.

SOLUTION: Put both uglify-js and uglify-es in the package.json file and remove
the part of the patch that changed package.json.

Aside from all this, the two newly-generated package-lock.json files are now added.
Testing showed that these two files cannot contain information about the
xcalar and xcalarsdk tar files.  npm fails with an obscure error when the
SHA in the package-lock.json file does not match the SHA of the actual tgz file
that it attempts to install in the new command #3 above.  Since this condition
occurs far more frequently than not, the only solution is to take xcalar and
xcalarsdk packages out of the checked in package-lock.json.

Two .npmrc files are also added; from the node/npm point of view the
services/expServer and assets/test/integrationTest/xcrpc are separate
subprojects that require their own files.

Change-Id: I9ca27cca56129fb9308b726fd5ebdb9e691ab0b5