XD-0000 Remove numRows parameter for ConfluentKafka
This change is to remove the numRows configuration for ConfluentKafka,
as the result is not accurate for now.

But we still keep the functionality to limit the num of rows(up to 20) when
previewing, so that to speed up the preview process.
When creating a dataset, we don't have numRows limitation, and import
all rows.

Change-Id: I5bdfcdd198be1c6ff6ec8ffbf2075f956c50e35b