ENG-5024 - add UserNotConfirmedException error code
This change causes the strategy to return a JSON structure that
includes more of the underlying SDK error state.  The JSON contains
three fields: message, code, and object.

message - a short description of the error
code - a text error code
object - the error itself when it either is not of type Error, does
         not have code and message fields, or cannot be converted to
         JSON.

This conversion occurs in the errMsgHandler routine.

Change-Id: I617348618efeba3abe2e977f2b094907a97ea231