ENG-501 - security features for the expServer in SaaS
In order to better secure Xcalar SaaS cluster, this change does
four things:

1) It gets the owner of the cluster from EC2 metadata (to get the
instance ID) and tags (to get the owner name).

2) It only allows two names to use expServer login on the cluster:
the owner's name and xdpadmin (in case we want to allow a private
Xcalar user later.)

3. It limits expServer requests to sessions whose username matches
the owner name.

4. It fixes broken cookie handling for SaaS login by correctly setting
the domain for the cookies created by the auth lambda.

Change-Id: I1f2c88a29a22a2b9645b1cf36cf14a70d13e368d