ENG-464 - Provide API endpoint to test cluster's health
To deal with the problem of determining if a cluster node
is functioning or not, this change adds a series of health
check endpoints to the expServer.  These endpoints are:

/service/healthCheck
/service/healthCheckUsrnode
/service/healthCheckMgmtd
/service/healthCheckJupyter

The healthCheck indicates that the expServer is running and
responding to requests.  The other three endpoints issue
requests of one sort or another to the usrnode (via xcrpc),
mgmtd (basic socket connection test -- due to time and the
fact that most native thrift calls are replaced by xcrpc), and
jupyter (http).

This strategy allows a simple http client to make a series of
calls that produce up/down accessments of Xcalar services.

Change-Id: I811906b51512b7a87ee72c854bbd43c9dc2501b9