ENG-5525 - SQLDF authentication fails for Xcalar SaaS
Changes in https://gerrit.int.xcalar.com/#/c/18226 for ENG-5396
altered the cookie output of the SaaS auth lambda in order to make
SDK authentication work.  Specifically, the auth lambda went from
returning a single cookie with the name connect.sid to returning
two with that name, one without a domain and one with the SAAS lambda
domain.

This causes problems for sqldf because it expects a single connect.sid
cookie, as with the on-prem deployment.

This change causes expServer login to return a single connect.sid
cookie in all cases.  It does this by filtering out the cookie
without a domain; the expServer modifies the other cookie to have
the same domain as what the http client should assign to the domain-
less cookie.

Change-Id: I3f4feea07f0c1d21abf0f26d52b7bf262f17eb61