ENG-6772 Fix grunt watch issue for SqlQueryHistoryPanel files
In grunt dev, tsc goes through all the ts files(in the src folder),
and is smart enough to prefix symbols with the correct namespace.

In grunt watch, only the file modified is present in the
build src(tswatchtmp), and tsc has no idea of the scope of
symbols because lack of context.
Ex. DynaTalbe would be treated as a global class,
if its definition file(DynaTable.ts) is not provided.

Change-Id: I8471fcbee71c9eec4ccb47c4f17468b8ee6c9274