ENG-6555 fix schema comparisons
My original code was checked into ts dir, but it was
duplicated in the src directory by another developer.
It could have just been moved. For some reason it was
copied over and recommitted a second time as the other
developer. So I am now deleting my original directory
and using this new duplicated directory and also fixed
some issues in the code.

Change-Id: Ib761a4e5ad444bcd488ff52ca43e3182854dc718