ENG-7373 xcalar operationalization
This includes the xcalar operationalization
code which largely consists of the xcalar_job.py
script that allows us to create, list, view
and delete xcalar jobs. It takes as input
a job json and creates a job or event on
a well known workbucket/eventprefix location.
This triggers an event that creates a schedule

Change-Id: I5701f737e1a9dc26e5b834b50ef9de0fb7813e52