ENG-7586 Hide old XD load and non cloud target in data mart
If it's cloud deployment, old XD load will be hide, and in export
panel only AWS Target will be available.
If it's not cloud deployment, old XD load and other connectors will be available.

We're using gDataMart = true to tell if it's cloud deployment or not,
it will only be defined on AWS VM now

Change-Id: Id221b0ad8d41145d1d588de44e2f939b24f09c20