ENG-7840 - A Xcalar authentication layer for terminalServer
The wetty package used by terminalServer is capable of invoking only
two programs: ssh and sshpass.   This change introduces a script called
ssh that can be invoked by wetty to allow it to authenticate with a
Xcalar cluster and then do other things.  (Currently, it just runs
htop.)

It previously used curl to perform authentication, but curl may not be
installed on customer systems.  It now uses a python script to perform
the authentication, employing packages found in xcalar-python36.

With the python authentication, it also saves the authentication tokens
into a Xcalar SDK-compatible cookiejar and sets the XCE_COOKIE_JAR_FILE
environment variable.  This should let the standard Xcalar SDK automatically
be able to work in an authenticated fashion.

It is the responsibility of any program run by the ssh script to delete
the cookie jar.

Change-Id: I189e204b1423a682c3a04cb980957e0878b158c6