ENG-8108 support for schema as a dict
In this version, the schema will arrive as a dict
{"rowpath" : <rowpath>, "columns" : <columns>}
instead of a list [name : ....]. This will be a
dict str which can then we converted to a json.
This is done because we have to use the rowpath
in the backend, automatically supplied by Kinesis
to tell us whether we have a top level array json
object in a source file. Using a rowpath helps to
make this N times faster where N is number of columns
compared to checking for top level array for each
columns' mapping. We had to send the new format to
XD because the schema is stored in KVS as a dict
and there is no way to avoid the schema being fetched
from KVS in schema_discover_load.py in
self._wb_kvs.lookup(schema_kvs_key) logic.

Change-Id: Ie24d6ee6275c5776a84e11f68f134e25859a0ac3