ENG-8377 Pass user_name to SchemaDiscoverApp
In order to access the load session, we need to pass both the
sessionName as well as the logged in user's user_name.

Otherwise, SchemaDiscoverApp will default to "admin", which works in our
internal testing, but not in AWS or customer deployment where the logged
in user is usually not "admin".

Change-Id: I797fd763b2367d721d2b092669101a0f25d2ddb1