ENG-8643 Copy Loader UDF into workbook UDF
In the new Load, we are using a UDF to populate the file list, schema,
and setup components. The UDF must be in a shared space because the
table load happens outside of the context of a project. However, for
operationalization, we need to move the shared UDF into the
workbook/udfs/ folder on download.

In the case of an SDK workbook-dataflow execution, the system will
simply look inside the workbook UDF folder for the UDF, and everything
will work.

Change-Id: I8f594315e2730454d14180045a31752ab6984b2f