ENG-8454 Error handler in thrift.js
We used to have a error handler in thrift.js for a edge case,
but this change get lost in a thrift version update in Dec 2017.
We don't hit that edge error for a while but according ENG-8454,
it comes back again.
Thus the error handler is added back.
As it's a thrift file change, there will be another corresponding
change on XCE side

Change-Id: I051b252e898cfe0b0e72054640d7ab76d175c947