ENG-9308 - http server error handling fixes
The expServer was previously attempting to handle errors at the
http.Server level (the highest level of the http stack that hosts
Express) using a handler for the 'error' event.  The handler
simply logged an error message saying "error on error handler".

There are two problems with this approach.  The http.Server object
does not have an 'error' event; the event is called 'clientError'
instead.  The defined error handler is not even being called.
Instead, a default 'clientError' event handler is being used that
does something like this:

server.on('clientError', (err, socket) => {
  socket.end('HTTP/1.1 400 Bad Request\r\n\r\n');
});

This change removes the existing 'error' handler and replaces it
with a new 'clientError' event handler.  The new handler does two
things.  It prints the error in the xcConsole log.  It then creates
a http 400 response that also includes the error stack.  Since the
Xcalar SDK client wants the body of the error message to be JSON,
the handler includes a function, replaceErrors, that can convert the
error object into the JSON object the SDK requires.

(cherry picked from commit cd1325208718641c7614d5abd5bf10669867ccaa)

Change-Id: I799e7615442d911ba571349a95c580e74430c716